requirements.txt
torch>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
PyYAML>=6.0
tqdm>=4.66.0

